+++ packages/types-codec/src/extended/BitVec.ts
@@ -41,13 +41,18 @@ function decodeBitVec (value?: AnyU8a): [number, Uint8Array] {
  */
 export class BitVec extends Raw {
   readonly #decodedLength: number;
+  readonly #isMsb?: boolean;
 
-  constructor (registry: Registry, value?: AnyU8a) {
+  // In lieu of having the Msb/Lsb identifiers passed through, we default to assuming
+  // we are dealing with Lsb, which is the default (as of writing) BitVec format used
+  // in the Polkadot code (this only affects the toHuman displays)
+  constructor (registry: Registry, value?: AnyU8a, isMsb = false) {
     const [decodedLength, u8a] = decodeBitVec(value);
 
     super(registry, u8a);
 
     this.#decodedLength = decodedLength;
+    this.#isMsb = isMsb;
   }
 
   /**
@@ -67,7 +72,12 @@ export class BitVec extends Raw {
   }
 
   public override toHuman (): string {
-    return `0b${[...this.toU8a(true)].map((d) => `00000000${d.toString(2)}`.slice(-8)).join('_')}`;
+    return `0b${
+      [...this.toU8a(true)]
+        .map((d) => `00000000${d.toString(2)}`.slice(-8))
+        .map((s) => this.#isMsb ? s : s.split('').reverse().join(''))
+        .join('_')
+    }`;
   }
 
   /**